Make `:username` optional configuration argument
When developing locally on Postgres and SQLite, it's common to rely on
the default username and have no password, which happens when creating
the database with `createdb mydatabase`. So, it would be convenient not
to have to pass the username. It seems that the database URL generator
is already handling the scenario of `:username` being `nil`.